% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{descApp}
\alias{descApp}
\title{descAPP for Descriptive statistics and Visualization}
\usage{
descApp()
}
\description{
descApp() function opens up an interactive shiny app which will allow
the user to easily calculate Summary Statistics, Summary Statistics by Group, Box plot,
Histogram, Q-Q plot and Shapiro-Wilk's test by uploading CSV file.
}
\details{
This app uses \code{descr} and \code{stby} functions of \code{summarytools}
package (Dominic Comtois, 2021) to calculate summary statistics and
summary statistics by group. \code{knitr} (Yihui Xie,2021) and \code{kableExtra}(Hao Zhu,2021) packages
were used to produce HTML tables. \code{shapiro.test}, \code{qqnorm} and \code{qqline} functions of
\code{stats}  package were used for Test of Homogeneity of variance and obtaining
Q-Q plot. \code{hist} and \code{boxplot} of package \code{graphics} were used
to obtain histogram and boxplot respectively. \code{ggqqplot} of package \code{ggpubr} (Alboukadel Kassambara,2020)
is also used to plot Q-Q plot in the app.
}
\examples{
\dontrun{
library(grapes)
grapes::descAPP()
}


}
\references{
\insertRef{Dominic_Comtois_2021}{grapes}

\insertRef{Hao_zhu_2021}{grapes}

\insertRef{Yihui_Xie_2021}{grapes}

\insertRef{R_2021}{grapes}

\insertRef{shiny_2021}{grapes}

\insertRef{sw_2021}{grapes}

\insertRef{dplyr_2021}{grapes}

\insertRef{ggpubr_2020}{grapes}

\insertRef{past_2018}{grapes}

\insertRef{magi_2020}{grapes}

\insertRef{gridG_2020}{grapes}
}
\keyword{box}
\keyword{by}
\keyword{descriptive}
\keyword{group}
\keyword{histogram}
\keyword{plot}
\keyword{q-q}
\keyword{statistics}
\keyword{summary}
