% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttApp}
\alias{ttApp}
\title{test and Paired t-test}
\usage{
ttApp()
}
\description{
ttApp() function opens up an interactive shiny app which will allow
user to easily perform one sample t-test, unpaired two sample t-test
unpaired two sample Welch t-test, paired t-test, test for homogeneity of variance (F-test),
and obtain plots like boxplot and paired plot by uploading CSV file.
}
\details{
Shiny App for t-test
}
\examples{
\dontrun{
library(grapes)
grapes::ttAPP()
}
}
\keyword{(F-test),}
\keyword{Testing}
\keyword{Welch}
\keyword{and}
\keyword{boxplot}
\keyword{homogeneity}
\keyword{of}
\keyword{one}
\keyword{paired}
\keyword{plot}
\keyword{sample}
\keyword{t-test}
\keyword{two}
\keyword{unpaired}
\keyword{variance}
