% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corrApp}
\alias{corrApp}
\title{corrAPP for correlation Analysis}
\usage{
corrApp()
}
\description{
corrApp() function opens up an interactive shiny app which will
allow the user to easily calculate Simple correlation, Correlation Matrix and obtain plots
like correlogram and scatterplot by uploading CSV file.
}
\details{
Shiny App for correlation Analysis
}
\examples{
\dontrun{
library(grapes)
grapes::corrAPP()
}
}
\keyword{Correlation}
\keyword{Correlogram}
\keyword{Matrix}
\keyword{Scatter}
\keyword{plot}
