% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{descApp}
\alias{descApp}
\title{descAPP for Descriptive statistics and Visualization}
\usage{
descApp()
}
\description{
descApp() function opens up an interactive shiny app which will allow
the user to easily calculate Summary Statistics, Summary Statistics by Group, Box plot,
Histogram, Q-Q plot and Shapiro-Wilk's test by uploading CSV file.
}
\details{
Shiny App for descriptive and exploratory analysis
}
\examples{
\dontrun{
library(grapes)
grapes::descAPP()
}
}
\keyword{box}
\keyword{by}
\keyword{descriptive}
\keyword{group}
\keyword{histogram}
\keyword{plot}
\keyword{q-q}
\keyword{statistics}
\keyword{summary}
